+++ src/tests/unit/tests/DetailsView/components/interactive-header.test.tsx
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { ExpandCollpaseLeftNavButtonProps } from 'common/components/left-nav-hamburger-button';
+import { ExpandCollpaseLeftNavButtonProps } from 'common/components/expand-collapse-left-nav-hamburger-button';
 import { DropdownClickHandler } from 'common/dropdown-click-handler';
 import { NamedFC, ReactFCWithDisplayName } from 'common/react/named-fc';
 import { DetailsViewPivotType } from 'common/types/details-view-pivot-type';
